pyproject.toml
[project]
name = "seminary-formation-platform"
version = "0.7.0"
description = "V7 backend for predictive formation, adaptive curriculum, outcomes, and class orchestration."
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.7",
    "fastapi>=0.111",
    "numpy>=1.26",
    "openai>=1.40",
    "pydantic>=2",
    "redis>=5",
    "sqlalchemy>=2",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "fakeredis>=2.20",
    "httpx>=0.27",
    "playwright>=1.45",
    "pytest>=8",
    "pytest-playwright>=0.5",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
